Change tests subfolders structure.
