+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/combine-root-definitions/file1.css
