+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/divide-classes/file2.css
