+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/filter-root-definitions/input.ts
