+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/ignore-comments/file.css
