+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/import-only-once/input.ts
