+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/multiple-import/input.ts
