+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/omit-root-definitions/file2.css
