+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/omit-root-definitions/input.ts
