+++ packages/ckeditor5-dev-build-tools/tests/plugins/splitCss/fixtures/single-import/input.ts
