+++ packages/ckeditor5-dev-build-tools/tests/plugins/translations/fixtures/pl.po
