+++ packages/ckeditor5-dev-build-tools/tests/plugins/translations/translations.test.ts
@@ -3,10 +3,12 @@
  * For licensing, see LICENSE.md.
  */
 
+import { join } from 'path';
 import { test, expect } from 'vitest';
 import { rollup, type OutputChunk, type RollupOutput } from 'rollup';
+import { verifyChunk } from '../../_utils/utils.js';
 
-import { translations, type RollupTranslationsOptions } from '../../src/index.js';
+import { translations, type RollupTranslationsOptions } from '../../../src/index.js';
 
 // eslint-disable-next-line max-len
 const ALL_POLISH_TRANSLATIONS = '\nexport default {"pl":{"dictionary":{"Hello world":"Witaj świecie","%0 files":["%0 plik","%0 pliki","%0 plików","%0 plików"]},getPluralForm(n){return (n==1 ? 0 : (n%10>=2 && n%10<=4) && (n%100<12 || n%100>14) ? 1 : n!=1 && (n%10>=0 && n%10<=1) || (n%10>=5 && n%10<=9) || (n%100>=12 && n%100<=14) ? 2 : 3);}}}';
@@ -23,9 +25,12 @@ const ENGLISH_TRANSLATIONS_FROM_ROOT = '\nexport default {"en":{"dictionary":{"H
 /**
  * Helper function for creating a bundle that won't be written to the file system.
  */
-async function generateBundle( options?: RollupTranslationsOptions, banner?: string ): Promise<RollupOutput['output']> {
+async function generateBundle(
+	options?: RollupTranslationsOptions,
+	banner?: string
+): Promise<RollupOutput['output']> {
 	const bundle = await rollup( {
-		input: './tests/plugins/fixtures/translations/input.js',
+		input: join( import.meta.dirname, './fixtures/input.js' ),
 		plugins: [
 			translations( options )
 		]
@@ -37,29 +42,14 @@ async function generateBundle( options?: RollupTranslationsOptions, banner?: str
 }
 
 /**
- * Helper function for validating a translation file.
- */
-function verifyOutputTranslations(
-	output: RollupOutput['output'],
-	filename: string,
-	translations: string
-): void {
-	const translation = output.find( output => output.name === filename );
-
-	expect( translation ).toBeDefined();
-	expect( translation!.type ).toBe( 'chunk' );
-	expect( ( translation as OutputChunk ).code ).toBe( translations );
-}
-
-/**
  * Test how the plugin behaves when no custom options are passed.
  */
 test( 'default options', async () => {
 	const output = await generateBundle();
 
-	verifyOutputTranslations( output, 'translations/pl.js', ALL_POLISH_TRANSLATIONS );
-	verifyOutputTranslations( output, 'translations/de.js', GERMAN_TRANSLATIONS_FROM_ROOT );
-	verifyOutputTranslations( output, 'translations/en.js', ENGLISH_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/pl.js', ALL_POLISH_TRANSLATIONS );
+	verifyChunk( output, 'translations/de.js', GERMAN_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/en.js', ENGLISH_TRANSLATIONS_FROM_ROOT );
 } );
 
 /**
@@ -69,9 +59,9 @@ test( 'banner', async () => {
 	const banner = 'BANNER';
 	const output = await generateBundle( undefined, banner );
 
-	verifyOutputTranslations( output, 'translations/pl.js', banner + ALL_POLISH_TRANSLATIONS );
-	verifyOutputTranslations( output, 'translations/de.js', banner + GERMAN_TRANSLATIONS_FROM_ROOT );
-	verifyOutputTranslations( output, 'translations/en.js', banner + ENGLISH_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/pl.js', banner + ALL_POLISH_TRANSLATIONS );
+	verifyChunk( output, 'translations/de.js', banner + GERMAN_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/en.js', banner + ENGLISH_TRANSLATIONS_FROM_ROOT );
 } );
 
 /**
@@ -79,12 +69,12 @@ test( 'banner', async () => {
  */
 test( 'source', async () => {
 	const output = await generateBundle( {
-		source: './tests/plugins/fixtures/translations/*.po'
+		source: join( import.meta.dirname, './fixtures/*.po' )
 	} );
 
-	verifyOutputTranslations( output, 'translations/pl.js', POLISH_TRANSLATIONS_FROM_ROOT );
-	verifyOutputTranslations( output, 'translations/de.js', GERMAN_TRANSLATIONS_FROM_ROOT );
-	verifyOutputTranslations( output, 'translations/en.js', ENGLISH_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/pl.js', POLISH_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/de.js', GERMAN_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'translations/en.js', ENGLISH_TRANSLATIONS_FROM_ROOT );
 } );
 
 /**
@@ -95,7 +85,7 @@ test( 'destination', async () => {
 		destination: 'languages'
 	} );
 
-	verifyOutputTranslations( output, 'languages/pl.js', ALL_POLISH_TRANSLATIONS );
-	verifyOutputTranslations( output, 'languages/de.js', GERMAN_TRANSLATIONS_FROM_ROOT );
-	verifyOutputTranslations( output, 'languages/en.js', ENGLISH_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'languages/pl.js', ALL_POLISH_TRANSLATIONS );
+	verifyChunk( output, 'languages/de.js', GERMAN_TRANSLATIONS_FROM_ROOT );
+	verifyChunk( output, 'languages/en.js', ENGLISH_TRANSLATIONS_FROM_ROOT );
 } );
